pyproject.toml
[project]
name = "sabormirandiano"
version = "2026.2.3"
description = "SaborMirandiano"
authors = [{ name = "Mario Hernandez", email = "mariofix@proton.me" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-security[babel,common,fsqla] (>=5.6.2,<6.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "flask-admin[images,sqlalchemy-with-utils,translation] (>=2.0.2,<3.0.0)",
    "flask-babel (>=4.0.0,<5.0.0)",
    "celery[redis,sqlalchemy] (>=5.6.0,<6.0.0)",
    "flask-mailman (>=1.1.1,<2.0.0)",
    "gunicorn",
    "dramatiq[redis,watch] (>=2.0.1,<3.0.0)",
    "wtforms-sqlalchemy (>=0.4.2,<0.5.0)",
    "black",
    "coverage",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-flask",
    "rich",
    "flask-debugtoolbar",
    "isort (>=7.0.0,<8.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "sentry-sdk[flask] (>=2.53.0,<3.0.0)",
    "khipu-tools (>=2025.1.0,<2026.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.coverage.run]
source = ["flask_merchants", "merchants", "app"]
omit = ["tests/*", "*/__init__.py", "*/migrations/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
show_missing = true
fail_under = 0

[tool.pytest.ini_options]
addopts = "-v --cov=flask_merchants --cov=merchants --cov=app --cov-report=xml --cov-report=term"
testpaths = ["tests"]
python_files = ["test*.py"]

[tool.black]
line-length = 119
target-version = ['py312']
